+++ src/syncback.rs
@@ -7,13 +7,13 @@ use std::collections::HashMap;
 const FORBIDDEN_CHARACTERS: [char; 1] = ['/'];
 
 #[cfg(windows)]
+const FORBIDDEN_CHARACTERS: [char; 9] = ['<', '>', ':', '"', '/', '\\', '|', '?', '*'];
+
+#[cfg(windows)]
 const FORBIDDEN_FILE_NAMES: [&str; 22] = [
 	"CON", "PRN", "AUX", "NUL", "COM1", "COM2", "COM3", "COM4", "COM5", "COM6", "COM7", "COM8", "COM9", "LPT1", "LPT2",
 	"LPT3", "LPT4", "LPT5", "LPT6", "LPT7", "LPT8", "LPT9",
 ];
-
-#[cfg(windows)]
-const FORBIDDEN_CHARACTERS: [char; 9] = ['<', '>', ':', '"', '/', '\\', '|', '?', '*'];
 
 pub fn verify_name(name: &str) -> Result<()> {
 	if name.is_empty() {